09-serverless/requirements.txt
numpy
Pillow
onnxruntime
torchvision